Initial work for julia 1.2 compatibility
Simple sub/pub with a local mosquitto server works.
Tests against the mock socket (mostly) pass.
Sometimes disconnect/keep-alive tests fail, so there appears to be some kind of race-condition
in the way disconnects are handled.
"Smoke" tests against `test.mosquitto.org` are failing.

+++ src/packets/disconnect.jl
@@ -1,5 +1,5 @@
 struct Disconnect <: Packet
     header::UInt8
 end
-Disconnect() = Disconnect(DISCONNECT)
+Disconnect() = Disconnect(UInt8(DISCONNECT))
 Base.show(io::IO, x::Disconnect) = print(io, "disconnect")
