+++ src/packets/ping.jl
@@ -1,11 +1,11 @@
 struct Pingreq <: Packet
     header::UInt8
 end
-Pingreq() = Pingreq(PINGREQ)
+Pingreq() = Pingreq(UInt8(PINGREQ))
 Base.show(io::IO, x::Pingreq) = print(io, "pingreq")
 
 struct Pingresp <: Packet
     header::UInt8
 end
-Pingresp() = Pingresp(PINGRESP)
+Pingresp() = Pingresp(UInt8(PINGRESP))
 Base.show(io::IO, x::Pingresp) = print(io, "pingresp")
